Move CloudServiceBase
