+++ src/api/PhotoCloudService.cs
@@ -1,5 +1,6 @@
 ﻿using Amazon.S3;
 using BDMS.Models;
+using BDMS.Services;
 using System.Text.RegularExpressions;
 
 namespace BDMS;
